Add methods to send a mailing to a designated list
Add configuration params to engage for mailings
Add method post_request to make it easy to post a specific command to engage.
Add method send_mailing to send a mailing out to a designated list or database
Add an errors file with exceptions
Add methods to retrieve the return_xml in string and hash form.
Add some specs to the engage_spec.rb file.